feat: create shared markdown service and fix object rendering
- Create centralized markdownService.js with shared rendering functions
- Update ChatSidebar.vue to use shared markdown service
- Fix [object Object] display issue in chat messages
- Update MarkdownPreview.vue to use shared service
- Remove code duplication across components
