+++ frontend/src/components/MarkdownPreview.vue
@@ -16,11 +16,10 @@
 
 <script setup>
 import { computed, onMounted, onUnmounted } from "vue";
-import { Marked } from "marked";
-import highlightjs from "highlight.js";
 import "highlight.js/styles/github.css";
 import "../styles/code-blocks.scss";
 import alertService from "../services/alertService";
+import { renderMarkdown, downloadCodeBlock } from "../services/markdownService";
 
 const props = defineProps({
   content: {
@@ -29,259 +28,25 @@ const props = defineProps({
   },
 });
 
-// Helper function to escape HTML for safe display
-function escapeHtml(unsafe) {
-  if (typeof unsafe !== 'string') {
-    return '';
-  }
-  return unsafe
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#039;");
-}
-
-/**
- * Deep content extractor - Robustly handles various content formats and edge cases
- * 
- * @param {*} input - The input content which could be anything (string, object, null, etc.)
- * @returns {string} - The extracted content as a string
- */
-function extractRealContent(input) {
-  // Case 1: Handle null/undefined
-  if (input === null || input === undefined) {
-    return '';
-  }
-  
-  // Case 2: Input is already a string
-  if (typeof input === 'string') {
-    // Check if it's a stringified JSON that needs parsing
-    try {
-      if ((input.trim().startsWith('{') && input.trim().endsWith('}')) || 
-          (input.trim().startsWith('[') && input.trim().endsWith(']'))) {
-        const parsed = JSON.parse(input);
-        // Recursively extract from the parsed object
-        return extractRealContent(parsed);
-      }
-    } catch (e) {
-      // Not valid JSON, just return the string
-    }
-    return input;
-  }
-  
-  // Case 3: Input is an object
-  if (typeof input === 'object') {
-    // Special handling for token objects with raw markdown
-    if (input.type === 'code') {
-      // If we have text content, just return it directly (don't add markdown fence)
-      if (input.text && typeof input.text === 'string') {
-        return input.text;
-      }
-    }
-    
-    // Priority property checking - check most likely properties first
-    const props = ['text', 'content', 'raw', 'value', 'markdown', 'md'];
-    for (const prop of props) {
-      if (input[prop] && typeof input[prop] === 'string') {
-        return input[prop];
-      }
-    }
-    
-    // Check if object has a reasonable toString implementation
-    if (input.toString && typeof input.toString === 'function') {
-      const str = input.toString();
-      if (str !== '[object Object]') {
-        return str;
-      }
-    }
-    
-    // Last resort: stringify the object
-    try {
-      return JSON.stringify(input, null, 2);
-    } catch (e) {
-      return '[Object]';
-    }
-  }
-  
-  // Case 4: Any other type - convert to string
-  return String(input);
-}
-
-// Create a new Marked instance with our options
-const marked = new Marked({
-  breaks: true,
-  gfm: true,
-  async: false,
-  // Add custom renderer for code blocks
-  renderer: {
-    code(code, language) {
-      let codeContent;
-      let codeLanguage = language || '';
-      
-      // Check for markdown code fences first and strip them if present
-      if (typeof code === 'string' && code.trim().startsWith('```')) {
-        const match = /^```([^\n]*)\n([\s\S]*?)```\s*$/g.exec(code);
-        if (match) {
-          if (match[1]) codeLanguage = match[1].trim();
-          codeContent = match[2];
-        } else {
-          codeContent = code;
-        }
-      }
-      // Handle case where code is a token object
-      else if (typeof code === 'object' && code !== null) {
-        // Extract language from token if available
-        if (code.lang) {
-          codeLanguage = code.lang;
-        }
-        
-        // Extract code content from token
-        codeContent = extractRealContent(code);
-        
-        // Check if the extracted content still has code fences
-        if (typeof codeContent === 'string' && codeContent.trim().startsWith('```')) {
-          const match = /^```([^\n]*)\n([\s\S]*?)```\s*$/g.exec(codeContent);
-          if (match) {
-            if (match[1]) codeLanguage = match[1].trim();
-            codeContent = match[2];
-          }
-        }
-      } else {
-        // Direct content - just make sure it's a string
-        codeContent = String(code || '');
-      }
-      
-      // Highlight the code
-      let highlightedCode;
-      try {
-        if (codeLanguage && highlightjs.getLanguage(codeLanguage)) {
-          highlightedCode = highlightjs.highlight(codeContent, { language: codeLanguage }).value;
-        } else {
-          highlightedCode = escapeHtml(codeContent);
-        }
-      } catch (error) {
-        console.error("Error highlighting code:", error);
-        highlightedCode = escapeHtml(codeContent);
-      }
-      
-      // Format display language name
-      const displayLanguage = codeLanguage 
-        ? codeLanguage.charAt(0).toUpperCase() + codeLanguage.slice(1)
-        : "Plain text";
-      
-      // Create filename for download
-      const timestamp = new Date().getTime();
-      const extension = codeLanguage || "txt";
-      const filename = `code-${timestamp}.${extension}`;
-      
-      // Return custom code block with download button
-      return `
-        <div class="code-block-wrapper">
-          <div class="code-header">
-            <span class="code-language">${displayLanguage}</span>
-            <button class="code-download-btn" title="Download code" data-code="${encodeURIComponent(codeContent)}" data-filename="${filename}" onclick="window.downloadCodeBlock(this)">
-              <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"></path>
-                <polyline points="7 10 12 15 17 10"></polyline>
-                <line x1="12" y1="15" x2="12" y2="3"></line>
-              </svg>
-            </button>
-          </div>
-          <pre><code class="hljs ${codeLanguage || ''}">${highlightedCode}</code></pre>
-        </div>
-      `;
-    }
-  }
+// Compute the rendered markdown
+const renderedMarkdown = computed(() => {
+  const html = renderMarkdown(props.content);
+  return html || '';
 });
 
-// Helper function to detect and fix code blocks in content prior to parsing
-function preprocessMarkdown(content) {
-  // Standard extraction first to handle nested objects
-  const extractedContent = extractRealContent(content);
-  
-  // Handle the case where the entire content is a stringified code token
-  if (typeof extractedContent === 'string' && 
-      extractedContent.trim().startsWith('{') && 
-      extractedContent.trim().endsWith('}')) {
-    try {
-      const possibleToken = JSON.parse(extractedContent);
-      if (possibleToken.type === 'code' && possibleToken.text) {
-        // Extract actual content and language for the renderer to handle
-        return possibleToken.text;
-      }
-    } catch (e) {
-      // Not a valid JSON token object, continue with regular content
-    }
-  }
-  
-  // Clean any triple backtick fences that might be in the extracted content
-  if (typeof extractedContent === 'string') {
-    // Remove markdown code block syntax if the entire content is a code block
-    const codeBlockMatch = /^\s*```([^\n]*)\n([\s\S]*?)```\s*$/g.exec(extractedContent);
-    if (codeBlockMatch && codeBlockMatch[2]) {
-      // The content is a code block; let the renderer handle it as regular code
-      return codeBlockMatch[2];
-    }
-  }
-  
-  return extractedContent;
-}
-
-// Function to render markdown content
-const getRenderedMarkdown = () => {
-  try {
-    if (!props.content) return '';
-    
-    // Pre-process to handle complex content structures
-    const processedContent = preprocessMarkdown(props.content);
-    
-    // Parse the markdown with our configured marked instance
-    return marked.parse(processedContent);
-  } catch (error) {
-    console.error("Error parsing markdown:", error);
-    
-    // Show error message, but also attempt to render the raw content
-    const rawContent = typeof props.content === 'string' 
-      ? props.content 
-      : JSON.stringify(props.content, null, 2);
-      
-    return `<p class="error">Error rendering markdown: ${error.message}</p>
-            <pre>${escapeHtml(rawContent)}</pre>`;
-  }
-};
-
-// Compute the rendered markdown
-const renderedMarkdown = computed(() => getRenderedMarkdown());
-
-// Function to download code blocks
-const downloadCodeBlock = (button) => {
-  try {
-    const code = decodeURIComponent(button.getAttribute("data-code"));
-    const filename = button.getAttribute("data-filename");
-
-    const blob = new Blob([code], { type: "text/plain" });
-    const url = URL.createObjectURL(blob);
-
-    const a = document.createElement("a");
-    a.href = url;
-    a.download = filename;
-    document.body.appendChild(a);
-    a.click();
-
-    document.body.removeChild(a);
-    URL.revokeObjectURL(url);
-
-    alertService.showAlert(`Downloading ${filename}`, "success", 3000);
-  } catch (error) {
-    console.error("Error downloading code:", error);
+// Handle download code blocks
+const handleDownloadCodeBlock = (button) => {
+  const result = downloadCodeBlock(button);
+  if (result.success) {
+    alertService.showAlert(`Downloading ${result.filename}`, "success", 3000);
+  } else {
     alertService.showAlert("Failed to download code", "error", 3000);
   }
 };
 
 onMounted(() => {
   // Add download function to window object
-  window.downloadCodeBlock = window.downloadCodeBlock || downloadCodeBlock;
+  window.downloadCodeBlock = window.downloadCodeBlock || handleDownloadCodeBlock;
 });
 
 onUnmounted(() => {
